requirements.txt
asgiref==3.5.2
Django==4.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-countries==7.3.2
django-crispy-forms==1.14.0
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
